Smart Neighbourhood
An android application that acts as a connecting bridge for people in a
community/neighbourhood. People of that neighbourhood can login and post
questions,answers,blogs etc... People can also conduct polls about the
decisions undertaken by the community as a whole.We have also
implemented a request service feature where people can request service
from a carpenter/plumber/painter etc.. These requests are processed and
sent to the particular service provider through an SMS using exotel
API's. An admin verifies the existence of a user and sends a
confirmation SMS to the user. We have also used openweathermap API to
get current weather which is reflected on the home page.The background
image of our home page change depends on the weather outside. We have
partially designed an ad algorithm where the ads on the people's
timeline change according to the posts of the people using data mining.